Rename a bunch of THERMOVR stuff to THERMOLAB.
