+++ .github/workflows/manual_events.yml
@@ -49,7 +49,7 @@ on:
         - "ERROR"
 
 env:
-  OGD_BQ_GAMES: '["AQUALAB", "BACTERIA", "BALLOON", "CRYSTAL", "CYCLE_CARBON", "CYCLE_NITROGEN", "CYCLE_WATER", "EARTHQUAKE", "JOWILDER", "PENGUINS", "LAKELAND", "MAGNET", "TRANSFORMATION_QUEST", "THERMOVR", "WAVES", "WIND"]'
+  OGD_BQ_GAMES: '["AQUALAB", "BACTERIA", "BALLOON", "CRYSTAL", "CYCLE_CARBON", "CYCLE_NITROGEN", "CYCLE_WATER", "EARTHQUAKE", "JOWILDER", "PENGUINS", "LAKELAND", "MAGNET", "TRANSFORMATION_QUEST", "THERMOLAB", "WAVES", "WIND"]'
 
 jobs:
   export_data:
