+++ src/ogd/games/THERMOLAB/detectors/__init__.py
