+++ src/ogd/games/THERMOLAB/schemas/THERMOVR.json.template
